Report Dump/Changes to General Ledger
Excel general ledger is a report we are creating behind the scenes, so that we can choose what information we are dumping to excel file. Currently just saves as xlsx file. Need to make it so it does not save over a file if button is pressed again. Need to check that the parameters are saved from previous report such that only data from report parameters is "dumped".