Add Check Formatting Options to AP and RD Check Stubs
I don't think that there is any way to get to the AP Check Stubs without getting a row count on one check above 41. That seems like entirely too much to keep on one page with a voided check (or even the whole page). I don't think that check is even necessary with the latest changes to check formatting. We have ways of dealing with check overflow through the check overflow options in check format settings. It might be a good idea to have the option to mail Reprint AP Check Stubs in the same way we can reprint any RD Check Stubs.