Use server environment variables for auth.
I kept the passwords we were using in code there for now so developers can test without setting their environment variables on the their local machine.
This uses environment variables stored in service instead of passwords stored in code. This means that only people with access to the service should be able to access these passwords instead of just anyone with access to our code.