Add function to work with Description parameters
Adds XSelected_PrintOnPageLabelFunction to print criteria at top of report. May need to add more definitions for parameters (each definition needs a select parameter, beginning parameter, ending parameter, description parameter for each of those, a singular of that parameter, plural of that parameter, the (first) index, and the (last) index.

I also added a function that changes the parameter name of the description parameter based on whether "pbooSearchSortParameters=false" is passed to the query string.